// src/types/enumWeekDays.ts
export enum MidweekDays {
    SEGUNDA = "Segunda-feira",
    TERCA = "Terça-feira",
    QUARTA = "Quarta-feira",
    QUINTA = "Quinta-feira",
    SEXTA = "Sexta-feira"
} 

export enum EndweekDays {
    SABADO = "Sábado",
    DOMINGO = "Domingo",
}

export enum Months {
    JANEIRO = "Janeiro",
    FEVEREIRO = "Fevereiro",
    MARCO = "Março",
    ABRIL = "Abril",
    MAIO = "Maio",
    JUNHO = "Junho",
    JULHO = "Julho",
    AGOSTO = "Agosto",
    SETEMBRO = "Setembro",
    OUTUBRO = "Outubro",
    NOVEMBRO = "Novembro",
    DEZEMBRO = "Dezembro",
  }
  